timeStamp as folder name added to make unique of each media in bucket
