+++ src/config/config.ts
@@ -21,7 +21,7 @@ const appConfiguration: IConfiguration = {
     name: get(envVars, 'AML_SERVICE_DB_NAME', 'postgres'),
     user: get(envVars, 'AML_SERVICE_DB_USER', 'postgres'),
   },
-  bucketName: get(envVars, 'BUCKET_NAME', 'aml-dev-otsv83nane'),
+  bucketName: get(envVars, 'BUCKET_NAME', 'dummyBucket'),
   presignedUrlExpiry: get(envVars, 'AWS_EXPIRY_TIME', 1800) as number,
   bulkUploadFolder: get(envVars, 'UPLOAD_FOLDER', 'upload'),
   templateFolder: get(envVars, 'TEMPLATE_FOLDER', 'template'),
