+++ src/controllers/media/mediaRead.ts
@@ -5,30 +5,29 @@ import { ResponseHandler } from '../../utils/responseHandler';
 import { amlError } from '../../types/amlError';
 import httpStatus from 'http-status';
 import { schemaValidation } from '../../services/validationService';
-import mediaUploadSchema from './mediaUploadSchema.json';
-import { uploadUrl } from '../../services/awsService';
-import { appConfiguration } from '../../config';
+import mediaReadSchema from './mediaReadSchema.json';
+import { getPresignedUrl } from '../../services/awsService';
+import path from 'path';
 import mime from 'mime-types';
 
 export const apiId = 'api.media.upload';
-const { mediaFolder } = appConfiguration;
 
-const getMediaUploadURL = async (req: Request, res: Response) => {
+const getMediaReadURL = async (req: Request, res: Response) => {
   const requestBody = _.get(req, 'body');
   const msgid = _.get(req, ['body', 'params', 'msgid']);
   const dataBody = _.get(req, 'body.request');
   const resmsgid = _.get(res, 'resmsgid');
 
-  const isRequestValid: Record<string, any> = schemaValidation(requestBody, mediaUploadSchema);
+  const isRequestValid: Record<string, any> = schemaValidation(requestBody, mediaReadSchema);
   if (!isRequestValid.isValid) {
-    const code = 'MEDIA_UPLOAD_INVALID_INPUT';
+    const code = 'MEDIA_READ_INVALID_INPUT';
     logger.error({ code, apiId, msgid, resmsgid, requestBody, message: isRequestValid.message });
     throw amlError(code, isRequestValid.message, 'BAD_REQUEST', httpStatus.BAD_REQUEST);
   }
 
   const signedUrls = await Promise.all(
     dataBody.map(async (file: any) => {
-      const getSignedUrl = await uploadUrl(mediaFolder, file.category, file.fileName);
+      const getSignedUrl = await getPresignedUrl(`${file.src}/${file.fileName}`);
       if (getSignedUrl.error) {
         const code = 'SERVER_ERROR';
         logger.error({ code, apiId, msgid, resmsgid, message: getSignedUrl.message });
@@ -36,7 +35,7 @@ const getMediaUploadURL = async (req: Request, res: Response) => {
       }
       const mediaMetaData = getMediaMetaData(file.fileName);
       return {
-        media: { fileName: file.fileName, src: `${mediaFolder}/${file.category}`, mimeType: mediaMetaData.mimeTye, mediaType: mediaMetaData.mediaType },
+        media: { fileName: file.fileName, src: file.src, mimeType: mediaMetaData.mimeTye, mediaType: mediaMetaData.mediaType },
         url: getSignedUrl.url,
         expiresInSec: getSignedUrl.expiresInSec,
       };
@@ -47,9 +46,10 @@ const getMediaUploadURL = async (req: Request, res: Response) => {
   ResponseHandler.successResponse(req, res, { status: httpStatus.OK, data: { message: 'success', signedUrls } });
 };
 
+export default getMediaReadURL;
+
 const getMediaMetaData = (fileName: string) => {
-  const fileMimeType = mime.lookup(fileName) || 'unknown';
+  const fileExtension = path.extname(fileName);
+  const fileMimeType = mime.lookup(fileExtension) || 'application/octet-stream';
   return { mimeTye: fileMimeType, mediaType: fileMimeType.split('/')[0] };
 };
-
-export default getMediaUploadURL;
