+++ src/controllers/media/mediaUploadSchema.json
